+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * Deprecation warnings
 * Starlette >=0.36 compatible
 
+### Removed
+* Spatial field
+
 ## [3.1](https://pypi.org/project/lupyne/3.1/) - 2023-11-22
 ### Changed
 * Python >=3.8 required